// src/api/sample-and-diagnosis-test/models/TestListEntry.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sample And Diagnosis Test List Api
 * Sample And Diagnosis Test List management for Web-In-Cloud system
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: your_email@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Condition } from './Condition';
import {
    ConditionFromJSON,
    ConditionFromJSONTyped,
    ConditionToJSON,
} from './Condition';

/**
 * 
 * @export
 * @interface TestListEntry
 */
export interface TestListEntry {
    /**
     * Unique id of the entry in this test list
     * @type {string}
     * @memberof TestListEntry
     */
    id: string;
    /**
     * Name of patient in test list
     * @type {string}
     * @memberof TestListEntry
     */
    name?: string;
    /**
     * Unique identifier of the patient known to Web-In-Cloud system
     * @type {string}
     * @memberof TestListEntry
     */
    patientId: string;
    /**
     * Timestamp since when the patient entered the test list
     * @type {Date}
     * @memberof TestListEntry
     */
    testSince: Date;
    /**
     * Estimated time of entering sample and diagnosis. Ignored on post.
     * @type {Date}
     * @memberof TestListEntry
     */
    estimatedStart?: Date;
    /**
     * Estimated duration of sample and diagnosis visit. If not provided then it will be computed based on condition and sample and diagnosis settings
     * @type {number}
     * @memberof TestListEntry
     */
    estimatedDurationMinutes: number;
    /**
     * 
     * @type {Condition}
     * @memberof TestListEntry
     */
    condition?: Condition;
}

/**
 * Check if a given object implements the TestListEntry interface.
 */
export function instanceOfTestListEntry(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "patientId" in value;
    isInstance = isInstance && "testSince" in value;
    isInstance = isInstance && "estimatedDurationMinutes" in value;

    return isInstance;
}

export function TestListEntryFromJSON(json: any): TestListEntry {
    return TestListEntryFromJSONTyped(json, false);
}

export function TestListEntryFromJSONTyped(json: any, ignoreDiscriminator: boolean): TestListEntry {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'patientId': json['patientId'],
        'testSince': (new Date(json['testSince'])),
        'estimatedStart': !exists(json, 'estimatedStart') ? undefined : (new Date(json['estimatedStart'])),
        'estimatedDurationMinutes': json['estimatedDurationMinutes'],
        'condition': !exists(json, 'condition') ? undefined : ConditionFromJSON(json['condition']),
    };
}

export function TestListEntryToJSON(value?: TestListEntry | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'patientId': value.patientId,
        'testSince': (value.testSince.toISOString()),
        'estimatedStart': value.estimatedStart === undefined ? undefined : (value.estimatedStart.toISOString()),
        'estimatedDurationMinutes': value.estimatedDurationMinutes,
        'condition': ConditionToJSON(value.condition),
    };
}